requirements.txt
attrs==21.3.0
aws-cdk.assertions==1.134.0
aws-cdk.assets==1.137.0
aws-cdk.aws-acmpca==1.137.0
aws-cdk.aws-applicationautoscaling==1.137.0
aws-cdk.aws-autoscaling-common==1.137.0
aws-cdk.aws-certificatemanager==1.137.0
aws-cdk.aws-cloudformation==1.137.0
aws-cdk.aws-cloudwatch==1.137.0
aws-cdk.aws-codeguruprofiler==1.137.0
aws-cdk.aws-codestarnotifications==1.137.0
aws-cdk.aws-ec2==1.137.0
aws-cdk.aws-ecr==1.137.0
aws-cdk.aws-ecr-assets==1.137.0
aws-cdk.aws-efs==1.137.0
aws-cdk.aws-events==1.137.0
aws-cdk.aws-iam==1.137.0
aws-cdk.aws-iot==1.137.0
aws-cdk.aws-kms==1.137.0
aws-cdk.aws-lambda==1.137.0
aws-cdk.aws-logs==1.137.0
aws-cdk.aws-opensearchservice==1.137.0
aws-cdk.aws-route53==1.137.0
aws-cdk.aws-s3==1.137.0
aws-cdk.aws-s3-assets==1.137.0
aws-cdk.aws-sam==1.137.0
aws-cdk.aws-secretsmanager==1.137.0
aws-cdk.aws-signer==1.137.0
aws-cdk.aws-sns==1.137.0
aws-cdk.aws-sns-subscriptions==1.134.0
aws-cdk.aws-sqs==1.137.0
aws-cdk.aws-ssm==1.137.0
aws-cdk.cloud-assembly-schema==1.137.0
aws-cdk.core==1.137.0
aws-cdk.custom-resources==1.137.0
aws-cdk.cx-api==1.137.0
aws-cdk.region-info==1.137.0
boto3==1.20.26
botocore==1.23.26
cattrs==1.9.0
certifi==2021.10.8
charset-normalizer==2.0.9
constructs==3.3.171
idna==3.3
jmespath==0.10.0
jsii==1.50.0
publication==0.0.3
python-dateutil==2.8.2
requests==2.26.0
requests-aws4auth==1.1.1
s3transfer==0.5.0
six==1.16.0
typing-extensions==4.0.1
urllib3==1.26.7
